bug fixing and adding missing library
there is bug in line 14. take note of it. and you have to install flask-cors. Note that after you install flask-cors, you have to restart your pc in order to let it work, otherwise you will get an error that this library has not been installed. or cors not available in your machine.